# Optimize codec initialization