--- a/tests/EdgeDB.Tests.Benchmarks/Program.cs
+++ b/tests/EdgeDB.Tests.Benchmarks/Program.cs
@@ -2,6 +2,6 @@ using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Running;
 using EdgeDB.Tests.Benchmarks;
 
-BenchmarkRunner.Run<TypeBuilderBenchmarks>();
+BenchmarkRunner.Run<FullExecuteBenchmark>();
 
 await Task.Delay(-1);